+++ conanfile.py
@@ -22,7 +22,7 @@ ConanBool = [True, False]
 
 
 logger = logging.getLogger(__name__)
-__version__: str = "0.130.9"
+__version__: str = "0.132.0"
 
 
 class QtGqlRecipe(ConanFile):