+++ qtgqlcodegen/templates/macros/deserialize_concrete_field.jinja.hpp
@@ -22,25 +22,33 @@ auto 👉proxy_field.name👈_typename  = 👉proxy_field.name👈_data.value("_
 👉iterate_type_condition(choice,type_cond, do_on_meets, loop)👈
 {% endfor %}
 {% elif proxy_field.type.is_model -%}
-    👉proxy_field.concrete.type.member_type👈 👉proxy_field.name👈_init_list;
+    {% if proxy_field.type.of_type.is_builtin_scalar %}
+    std::vector<👉proxy_field.type.of_type.type_name()👈> 👉proxy_field.name👈_init_vec;
     for (const auto& node: data.value("👉proxy_field.name👈").toArray()){
-    {% if proxy_field.type.is_model.of_type.is_queried_object_type %}
-        👉proxy_field.name👈_init_list.append(👉 proxy_field.type.is_model.of_type.is_queried_object_type.deserializer_name 👈(node.toObject(), 👉operation_pointer👈));
-    {% elif proxy_field.type.is_model.of_type.is_queried_union or proxy_field.type.is_model.of_type.is_queried_interface %}
-        auto node_data = node.toObject();
-        auto 👉proxy_field.name👈_typename = node_data.value("__typename").toString();
-        {%set type_cond -%}👉proxy_field.name👈_typename{% endset -%}
-        {% for choice in proxy_field.type.of_type.choices -%}
-        {% set do_on_meets -%}
-        👉proxy_field.name👈_init_list.append(👉choice.deserializer_name👈(node_data, 👉operation_pointer👈) 👉 setter_end 👈);
-        {% endset -%}
-        👉iterate_type_condition(choice,type_cond, do_on_meets, loop)👈
-        {% endfor %}
+        👉proxy_field.name👈_init_vec.push_back(node.👉 proxy_field.type.of_type.from_json_convertor 👈);
+    }
+    👉 setter_name 👈(std::make_shared<👉proxy_field.concrete.type.type_name()👈>(nullptr, 👉proxy_field.name👈_init_vec) 👉 setter_end 👈);
     {% else %}
-    throw qtgql::exceptions::NotImplementedError({"can't deserialize model of 👉proxy_field.type.of_type.__class__👈"});
+        👉proxy_field.concrete.type.member_type👈 👉proxy_field.name👈_init_vec;
+        for (const auto& node: data.value("👉proxy_field.name👈").toArray()){
+        {% if proxy_field.type.is_model.of_type.is_queried_object_type %}
+            👉proxy_field.name👈_init_vec.push_back(👉 proxy_field.type.of_type.is_queried_object_type.deserializer_name 👈(node.toObject(), 👉operation_pointer👈));
+        {% elif proxy_field.type.is_model.of_type.is_queried_union or proxy_field.type.is_model.of_type.is_queried_interface %}
+            auto node_data = node.toObject();
+            auto 👉proxy_field.name👈_typename = node_data.value("__typename").toString();
+            {%set type_cond -%}👉proxy_field.name👈_typename{% endset -%}
+            {% for choice in proxy_field.type.of_type.choices -%}
+            {% set do_on_meets -%}
+            👉proxy_field.name👈_init_vec.push_back(👉choice.deserializer_name👈(node_data, 👉operation_pointer👈) 👉 setter_end 👈);
+            {% endset -%}
+            👉iterate_type_condition(choice,type_cond, do_on_meets, loop)👈
+            {% endfor %}
+        {% else %}
+        throw qtgql::exceptions::NotImplementedError({"can't deserialize model of 👉proxy_field.type.of_type.__class__👈"});
+        {% endif %}
+        };
+        👉 setter_name 👈(👉proxy_field.name👈_init_vec 👉 setter_end 👈);
     {% endif %}
-    };
-    👉 setter_name 👈(👉proxy_field.name👈_init_list👉 setter_end 👈);
 {% elif proxy_field.type.is_builtin_scalar -%}
     {% if proxy_field.type.is_void -%}
     /* deliberately empty */