+++ qtgqlcodegen/types.py
@@ -45,6 +45,10 @@ class QtGqlTypeABC(ABC):
         return None
 
     @property
+    def is_input_list(self) -> QtGqlInputList | None:
+        return None
+
+    @property
     def is_enum(self) -> QtGqlEnumDefinition | None:
         return None
 
@@ -156,21 +160,53 @@ class QtGqlList(QtGqlTypeABC):
         return self
 
     @property
+    def needs_proxy_model(self) -> bool:
+        """Model of scalars / enums.
+
+        have no further selections hence they are shared across the
+        schema.
+        """
+        if self.of_type.is_builtin_scalar or self.of_type.is_enum:
+            return True
+        return False
+
+    @property
     def member_type(self) -> str:
-        return f"QList<{self.of_type.member_type}>"
+        if self.of_type.is_builtin_scalar:
+            # scalars has no underlying fields hence they don't need to be
+            # proxied. So each proxy would get an instance to the model.
+            return f"std::shared_ptr<{self.type_name()}>"
+
+        return f"std::vector<{self.of_type.member_type}>"
 
     def type_name(self) -> str:
+        if bs := self.of_type.is_builtin_scalar:
+            return f"{QtGqlTypes.ListModelABC.name}<{bs.member_type}>"
         raise NotImplementedError(
-            "models have no valid type for schema concretes, call member_type",
+            "complex models have no valid type for schema concretes, call member_type",
         )
 
     @property
     def property_type(self) -> str:
         if self.of_type.is_queried_object_type or self.of_type.is_queried_interface:
-            return f"qtgql::bases::ListModelABC<{self.of_type.type_name()}> *"
+            return f"{QtGqlTypes.ListModelABC.name}<{self.of_type.property_type}> *"
         if self.of_type.is_queried_union:
-            return f"qtgql::bases::ListModelABC<{self.of_type.is_queried_union.type_name()}> *"
+            return f"{QtGqlTypes.ListModelABC.name}<{self.of_type.property_type}> *"
+        if bs := self.of_type.is_builtin_scalar:
+            return f"{QtGqlTypes.ListModelABC.name}<{bs.member_type}> *"
         raise NotImplementedError
+
+
+@define
+class QtGqlInputList(QtGqlTypeABC):
+    of_type: QtGqlTypeABC
+
+    @property
+    def is_input_list(self) -> QtGqlInputList | None:
+        return self
+
+    def type_name(self) -> str:
+        return f"std::list<{self.of_type.type_name()}>"
 
 
 @define
@@ -219,6 +255,10 @@ class BuiltinScalar(QtGqlTypeABC):
     @property
     def is_void(self) -> bool:
         return self is BuiltinScalars.VOID
+
+    @property
+    def property_type(self) -> str:
+        return self.type_name()
 
     def json_repr(self, attr_name: str | None = None) -> str:
         return f"{attr_name}"