+++ tests/test_codegen/schemas/__init__.py
@@ -3,6 +3,8 @@ from . import (
     input_type,
     list_of_interface,
     list_of_non_node,
+    list_of_scalar,
+    list_of_scalar_argument,
     list_of_union,
     mutation_schema,
     node_interface_field,
@@ -64,4 +66,6 @@ __all__ = [  # noqa: PLE0604
     subscription_schema,
     node_interface_field,
     list_of_interface,
+    list_of_scalar,
+    list_of_scalar_argument,
 ]