+++ CHANGELOG.md
@@ -1,6 +1,27 @@
 CHANGELOG
 =========
 
+0.133.0 - 2023-08-03
+--------------------
+
+This release adds support for list of scalars argument.
+
+Contributed by [ניר](https://github.com/nrbnlulu) via [PR #349](https://github.com/qtgql/qtgql/pull/349/)
+
+
+0.132.0 - 2023-08-02
+--------------------
+
+This release adds support for list of scalars.
+
+### Breaking Changes
+
+- Models now expose `data` role instead of `qtObject` role since
+it can also be an `int` for example.
+
+Contributed by [ניר](https://github.com/nrbnlulu) via [PR #348](https://github.com/qtgql/qtgql/pull/348/)
+
+
 0.131.0 - 2023-07-31
 --------------------